+++ src/System.Windows.Forms/src/System/Windows/Forms/ImageList.cs
@@ -904,7 +904,7 @@ namespace System.Windows.Forms {
             internal NativeImageList(IntPtr himl) {
                 this.himl = himl;
 #if DEBUG
-                new EnvironmentPermission(PermissionState.Unrestricted).Assert();
+                new EnvironmentPermission((int)PermissionState.Unrestricted).Assert();
                 try {
                     callStack = Environment.StackTrace;
                 }
